Removed several @ts-ignore

+++ commands/moderation/purge.ts
@@ -18,7 +18,11 @@ export const Purge: Command = {
     },
   ],
   run: async (client: Client, interaction: CommandInteraction) => {
-    // @ts-ignore
+    if (!interaction.channel?.isDMBased){
+      interaction.followUp({ content: "An error has occurred" });
+      return;
+    }
+    // @ts-expect-error
     const amount = interaction.options.getInteger('number')
     if (amount > 100) 
       return interaction.followUp({ content : `You can only purge 100 messages at a time.`, ephemeral: true });
@@ -28,9 +32,10 @@ export const Purge: Command = {
     const filtered = messages?.filter(
       (msg) => Date.now() - msg.createdTimestamp < ms("14 days")
     );
-    // @ts-ignore
+    if (filtered === undefined)
+      return interaction.followUp({ content: 'An error has occured', ephemeral: true });
     await interaction.followUp({ content : `Purged ${filtered.size - 1} messages.`, ephemeral: true });
-    // @ts-ignore
+    // @ts-expect-error
     await interaction.channel?.bulkDelete(filtered)
   },
 };
